created sign up, sign in and logout

+++ base/views.py
@@ -1,6 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from django.db.models import Q
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm  
+from django.contrib.auth.models import User
 from .forms import RoomForm
 
 # Create your views here.
@@ -10,6 +16,54 @@ from .forms import RoomForm
 #    {'id':2, 'name':'Code'},
 #    {'id':3, 'name':'Together'},
 # ]
+
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User doesn't exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+             messages.error(request, "User or Password doesn't exist")
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Registration not successful')   
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def home(request):
@@ -22,10 +76,12 @@ def home(request):
         )
 
     topics = Topic.objects.all()
+    room_count = rooms.count()
 
-    context = {'rooms': rooms,'topics': topics}
+    context = {'rooms': rooms,'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
     
+
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
@@ -33,6 +89,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -45,9 +102,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not a user!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -59,8 +121,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not a user!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
